+++ form.css
@@ -3,8 +3,8 @@
 
     height: 100%;
     border:1px solid black;
-    padding: 10px;
-    margin: 10px;
+    padding: 5px;
+    margin: 5px;
     float: left;
 }
 
@@ -12,26 +12,27 @@
     width: 20%;
     height: 100%;
     border:1px solid black;
-    margin-right:10px;
+    margin-right:  10px;
     float: left;
 }
 
 .form-field {
-	background-color: red;
+	/*background-color: red; */
 }
 
 input {
-	margin-left: 1em;
+	margin-left: 0.5em;
 }
 .Group {
-	background-color: yellow;
-	padding-left: 2em;
+	background-color: white;
+	/* padding-left: 2em; */
+	border: 1px solid black;
 }
 
 
 .VerticalLayout {
 	padding-left: 1em;
-	background-color: green;
+	/* background-color: green; */
 }
 
 
@@ -39,6 +40,6 @@ input {
 	padding-top: 1em;
 	padding-bottom: 1em;
 	height: 100%;
-	border: 2px solid black;
-	background-color: blue;
+	/*border: 2px solid black;
+	 background-color: blue; */
 }
